drm: panel: lg-sw43408: Add lab regulator SC notify support
Lab regulator requires the clients to register for the short circuit
notification sent via REGULATOR_EVENT_OVER_CURRENT.

They should try to enable the regulator on getting this event. If it
doesn't get enabled, they need to assume that short circuit is permanent
and disable themselves.

Add this support to the panel driver, who is a client.

Signed-off-by: Sumit Semwal <sumit.semwal@linaro.org>